Moved CLI into module
